Fixed a Tile Bug and Began Particle Work
I fixed a bug where the tiles were appearing in the wrong location in the scene, and I did some work on a particle effect.  I hope to apply this particle effect to the "Power Crystals" later in development.  For now, it is an interesting exercise in types of classes and drawing to the canvas.